Changed FS ok to exactly after load
Previously FS would say ok to MK3, MK3 would wait 1.5s then do C0 assuming the completion of steps to bondtech after FSensor  triggered. Wastes time and if value was increase for a custom setup it wouldn't be enough time.
No the ok from to MK3 is sent once the steps to bontech are complete, no delay on MK3 saving a few 100ms on stock and accommodating for custom lengths.